feat : updated datatypes for entities

+++ DepositReport.Domain/Entities/FileTypes.cs
@@ -2,7 +2,7 @@
 {
     public class FileTypes
     {
-        public int FileTypeID { get; set; }
+        public byte FileTypeID { get; set; }
         public string FileTypeName { get; set; }
         public string SourcePath { get; set; }
         public string Description { get; set; }
